fix(ResourceService): 将评论点赞数转换为整型
修复评论点赞数类型不匹配问题，将getCommentLikes返回的数值显式转换为int类型，避免潜在的比较错误

+++ src/main/java/com/neon/service/ResourceService.java
@@ -185,17 +185,17 @@ public class ResourceService {
         }
 
         for (Comment topComment : topLevelComments) {
-            topComment.setLikes(getCommentLikes(topComment.getId()));
+            topComment.setLikes(getCommentLikes(topComment.getId()).intValue());
             
             List<Comment> replies = repliesMap.get(topComment.getId());
             if (replies != null) {
                 for (Comment reply : replies) {
-                    reply.setLikes(getCommentLikes(reply.getId()));
+                    reply.setLikes(getCommentLikes(reply.getId()).intValue());
                     
                     List<Comment> nestedReplies = repliesMap.get(reply.getId());
                     if (nestedReplies != null) {
                         for (Comment nestedReply : nestedReplies) {
-                            nestedReply.setLikes(getCommentLikes(nestedReply.getId()));
+                            nestedReply.setLikes(getCommentLikes(nestedReply.getId()).intValue());
                         }
                         nestedReplies.sort((a, b) -> {
                             if (!b.getLikes().equals(a.getLikes())) {
